added read/write access param

+++ api/controllers/projectController.ts
@@ -10,7 +10,23 @@ exports.getProject = async function (req, res) {
       });
       return;
     }
-    res.json(req.project);
+    if (req.writeAccess) {
+      res.json({
+        id: req.project.id,
+        name: req.project.name,
+        token: req.project.token,
+        readToken: req.project.readToken,
+        documentIds: req.project.documentIds,
+        owner: req.project.owner,
+      });
+    } else {
+      res.json({
+        id: req.project.id,
+        name: req.project.name,
+        documentIds: req.project.documentIds,
+        owner: req.project.owner,
+      });
+    }
   } catch (err) {
     console.log(err);
     res.send(err);
@@ -18,6 +34,18 @@ exports.getProject = async function (req, res) {
 };
 exports.updateProject = async function (req, res) {
   try {
+    if (!req.project) {
+      res.status(404).send({
+        status: "Project not found",
+      });
+      return;
+    }
+    if (!req.writeAccess) {
+      res.status(403).send({
+        status: "Forbidden",
+      });
+      return;
+    }
     let newProj = new Project(req.body);
     let status = await Project.updateOne(
       {
@@ -37,6 +65,18 @@ exports.updateProject = async function (req, res) {
 };
 exports.deleteProject = async function (req, res) {
   try {
+    if (!req.project) {
+      res.status(404).send({
+        status: "Project not found",
+      });
+      return;
+    }
+    if (!req.writeAccess) {
+      res.status(403).send({
+        status: "Forbidden",
+      });
+      return;
+    }
     let status = await Project.deleteOne({
       id: req.project.id,
     }).exec();
@@ -52,6 +92,7 @@ exports.addProject = async function (req, res) {
     let newProj = new Project(req.body);
     newProj.id = randomUUID();
     newProj.token = Buffer.from(randomUUID()).toString("base64");
+    newProj.readToken = Buffer.from(randomUUID()).toString("base64");
     res.json(await newProj.save());
   } catch (err) {
     console.log(err);
